Added admin condition to restricted platfrom routes and fixed corresponding tests

+++ server/controller/platformsController.js
@@ -6,43 +6,64 @@ const getPlatforms = async (req, res) => {
 };
 
 const createPlatforms = async (req, res, next) => {
-  const newPlatform = await Platform.create(req.body);
-  res.json(newPlatform);
+  const { admin } = req.userInfo;
+  if (!admin) {
+    const error = new Error("You wish");
+    error.code = 401;
+    next(error);
+  } else {
+    const newPlatform = await Platform.create(req.body);
+    res.json(newPlatform);
+  }
 };
 
 const updatePlatforms = async (req, res, next) => {
-  const { idPlatform } = req.params;
-  const updatedPlatform = await Platform.findByIdAndUpdate(
-    idPlatform,
-    req.body,
-    {
-      new: true,
-    }
-  );
-  if (updatedPlatform) {
-    res.json(updatedPlatform);
-  } else {
-    const error = new Error("Platform not found");
-    error.code = 404;
+  const { admin } = req.userInfo;
+  if (!admin) {
+    const error = new Error("You wish");
+    error.code = 401;
     next(error);
+  } else {
+    const { idPlatform } = req.params;
+    const updatedPlatform = await Platform.findByIdAndUpdate(
+      idPlatform,
+      req.body,
+      {
+        new: true,
+      }
+    );
+    if (updatedPlatform) {
+      res.json(updatedPlatform);
+    } else {
+      const error = new Error("Platform not found");
+      error.code = 404;
+      next(error);
+    }
   }
 };
 
 const deletePlatforms = async (req, res, next) => {
-  try {
-    const { idPlatform } = req.params;
-    const deletedPlatform = await Platform.findByIdAndDelete(idPlatform);
-    if (deletedPlatform) {
-      res.json(deletePlatforms);
-    } else {
-      const error = new Error("Not found");
-      error.code = 404;
+  const { admin } = req.userInfo;
+  if (!admin) {
+    const error = new Error("You wish");
+    error.code = 401;
+    next(error);
+  } else {
+    try {
+      const { idPlatform } = req.params;
+      const deletedPlatform = await Platform.findByIdAndDelete(idPlatform);
+      if (deletedPlatform) {
+        res.json(deletePlatforms);
+      } else {
+        const error = new Error("Not found");
+        error.code = 404;
+        next(error);
+      }
+    } catch (error) {
+      error.code = 400;
+      error.message = "Cannot delete sorry :(";
       next(error);
     }
-  } catch (error) {
-    error.code = 400;
-    error.message = "Cannot delete sorry :(";
-    next(error);
   }
 };
